+++ src/pages/Homepage.tsx
@@ -1,8 +1,7 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { TripImageObject } from "../interfaces/SharedInterfaces"
 import ImageUpload from "../components/ImageUpload";
 import InteractiveMapWrapper from "../components/InteractiveMapWrapper";
-import { getGeoCodeByString } from "../api/Geocode/GeocodeApi";
 
 const Homepage = () => {
     const [imageMetaData, setImageMetaData] = useState<TripImageObject[]>([]);
